content/base/test/test_bug638112.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=638112
-->
<head>
  <title>Test for Bug 638112</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=638112">Mozilla Bug 638112</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="text/javascript">

/** Test for Bug 638112 **/

function run_test() {
  var req = new XMLHttpRequest();
  req.open("GET", "bug638112.sjs", false);
  req.send(null);
  var statusText = req.statusText;

  is(statusText, "Information Sans-Autorit", "");

  SimpleTest.finish();
}


SimpleTest.waitForExplicitFinish();

addLoadEvent(run_test);

</script>
</pre>
</body>
</html>
